ui updates and balance is now Double
changed the add payment input field and updated the ui

+++ app/src/main/java/com/example/roomy/db/BalanceRepository.kt
@@ -56,7 +56,7 @@ class BalanceRepository {
 
     // Balance function
     // add Balance when there is a new payment
-    suspend fun addBalance(groupId: Int, owedBy: String, owedTo: String, amount: Int): Balance? {
+    suspend fun addBalance(groupId: Int, owedBy: String, owedTo: String, amount: Double): Balance? {
 
         var updatedBalance: Balance? = null
         try {
